requirements.txt
# requirements.txt
pandas
scikit-learn
streamlit
google-genai
nltk